requirements.txt
requests
streamlit
pyyaml
python-dotenv